Senty: Add IP test package
Add a new IP test package that will hold all the IP tests,
this package is used to cover all the IP functionality.
In this change iperf_test, netperf_test, ping_test has been
modified to be used by IP package.
Also in this change basic_test and traffic_test has been modified
so they can be used as objects and not executables.

Signed-off-by: Erez Alfasi <erez8272@gmail.com>